requirements.txt
# Core frameworks
#torch>=2.0.0
torchvision>=0.15.0
lightning>=2.0.0
hydra-core>=1.3.0

# Model and tokenization
transformers>=4.30.0
tokenizers>=0.13.0

# Logging and monitoring
wandb>=0.15.0
tensorboard>=2.12.0

# Metrics and utilities
torchmetrics>=0.11.0
einops>=0.6.0
omegaconf>=2.3.0

# Data processing
pandas>=1.5.0
numpy>=1.21.0
Pillow>=9.0.0
opencv-python>=4.7.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Progress bars and utilities
rich>=13.0.0
tqdm>=4.64.0

# Optional: for model summary
torchsummary>=1.5.1

# Optional: for better performance
accelerate>=0.20.0